bot/requirements.txt
pinecone-client
openai
tiktoken
nest_asyncio
# os
# pinecone
fastapi~=0.103.1
langchain==0.0.305
uvicorn~=0.23.2